refactor: view controller로 분리
